requirements.txt
aiofiles>=23.1.0
fastapi>=0.95.0
faster-whisper>=0.3.0
ffmpeg-python>=0.2.0
loguru>=0.6.0
pyannote-audio==2.1.1
pydantic>=1.10.7
python-dotenv>=1.0.0
python-multipart>=0.0.6
scikit-learn>=1.2.2
uvicorn>=0.21.1
yt-dlp>=2023.3.4
